refactorized testsuite
