+++ extension.neon
@@ -1,5 +1,6 @@
 parameters:
 	ignoreErrors:
 		# Can't use the PHPUnit extension on PHP 5.6, ignore Mocks false positive
+		- '#Access to an undefined property PHPUnit_Framework_MockObject_Invocation#'
 		- '#should return PHPUnit_Framework_MockObject_MockObject#'
 		- '#PHPUnit_Framework_MockObject_MockObject given#'
