+++ tests/CircuitBreakerTestCase.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Tests\PrestaShop\CircuitBreaker;
+
+use PrestaShop\CircuitBreaker\Clients\GuzzleClient;
+use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Handler\MockHandler;
+use PHPUnit\Framework\TestCase;
+use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\HandlerStack;
+use GuzzleHttp\Psr7\Request;
+
+/**
+ * Helper to get a fake Guzzle client.
+ */
+abstract class CircuitBreakerTestCase extends TestCase
+{
+    /**
+     * Returns an instance of Client able to emulate
+     * available and not available services.
+     *
+     * @return GuzzleClient
+     */
+    protected function getTestClient()
+    {
+        $mock = new MockHandler([
+            new RequestException('Service unavailable', new Request('GET', 'test')),
+            new RequestException('Service unavailable', new Request('GET', 'test')),
+            new Response(200, [], '{"hello": "world"}'),
+        ]);
+
+        $handler = HandlerStack::create($mock);
+
+        return new GuzzleClient(['handler' => $handler]);
+    }
+}
