+++ tests/CircuitBreakerTests.php
@@ -2,30 +2,27 @@
 
 namespace Tests\PrestaShop\CircuitBreaker;
 
-use GuzzleHttp\Exception\RequestException;
-use GuzzleHttp\Handler\MockHandler;
-use GuzzleHttp\HandlerStack;
-use GuzzleHttp\Psr7\Request;
-use GuzzleHttp\Psr7\Response;
-use PHPUnit\Framework\TestCase;
-use PrestaShop\CircuitBreaker\Clients\GuzzleClient;
-use PrestaShop\CircuitBreaker\Places\ClosedPlace;
-use PrestaShop\CircuitBreaker\Places\HalfOpenPlace;
-use PrestaShop\CircuitBreaker\Places\OpenPlace;
+use Symfony\Component\EventDispatcher\EventDispatcher;
+use PrestaShop\CircuitBreaker\Storages\SymfonyCache;
+use PrestaShop\CircuitBreaker\SymfonyCircuitBreaker;
 use PrestaShop\CircuitBreaker\SimpleCircuitBreaker;
-use PrestaShop\CircuitBreaker\States;
+use PrestaShop\CircuitBreaker\Places\HalfOpenPlace;
+use PrestaShop\CircuitBreaker\Places\ClosedPlace;
+use PrestaShop\CircuitBreaker\Systems\MainSystem;
+use PrestaShop\CircuitBreaker\Places\OpenPlace;
+use Symfony\Component\Cache\Simple\ArrayCache;
 
-class SimpleCircuitBreakerTest extends TestCase
+class CircuitBreakerTests extends CircuitBreakerTestCase
 {
     /**
      * When we use the circuit breaker on unreachable service
      * the fallback response is used.
+     *
+     * @dataProvider getCircuitBreakers
      */
-    public function testCircuitBreakerIsInClosedStateAtStart()
+    public function testCircuitBreakerIsInClosedStateAtStart($circuitBreaker)
     {
-        $circuitBreaker = $this->createCircuitBreaker();
-
-        $this->assertSame(States::CLOSED_STATE, $circuitBreaker->getState());
+        $this->assertSame('CLOSED', $circuitBreaker->getState());
 
         $this->assertSame(
             '{}',
@@ -41,14 +38,14 @@ class SimpleCircuitBreakerTest extends TestCase
      * is opened. This time no calls to the services are done.
      *
      * @depends testCircuitBreakerIsInClosedStateAtStart
+     * @dataProvider getCircuitBreakers
      */
-    public function testCircuitBreakerWillBeOpenedInCaseOfFailures()
+    public function testCircuitBreakerWillBeOpenedInCaseOfFailures($circuitBreaker)
     {
-        $circuitBreaker = $this->createCircuitBreaker();
         // CLOSED
         $circuitBreaker->call('https://httpbin.org/get/foo', $this->createFallbackResponse());
 
-        $this->assertSame(States::OPEN_STATE, $circuitBreaker->getState());
+        $this->assertSame('OPENED', $circuitBreaker->getState());
         $this->assertSame(
             '{}',
             $circuitBreaker->call(
@@ -65,10 +62,10 @@ class SimpleCircuitBreakerTest extends TestCase
      *
      * @depends testCircuitBreakerIsInClosedStateAtStart
      * @depends testCircuitBreakerWillBeOpenedInCaseOfFailures
+     * @dataProvider getCircuitBreakers
      */
-    public function testAfterTheThresholdTheCircuitBreakerMovesInHalfOpenState()
+    public function testAfterTheThresholdTheCircuitBreakerMovesInHalfOpenState($circuitBreaker)
     {
-        $circuitBreaker = $this->createCircuitBreaker();
         // CLOSED
         $circuitBreaker->call('https://httpbin.org/get/foo', $this->createFallbackResponse());
         // OPEN
@@ -83,7 +80,7 @@ class SimpleCircuitBreakerTest extends TestCase
                 $this->createFallbackResponse()
             )
         );
-        $this->assertSame(States::HALF_OPEN_STATE, $circuitBreaker->getState());
+        $this->assertSame('HALF_OPENED', $circuitBreaker->getState());
         $this->assertTrue($circuitBreaker->isHalfOpened());
     }
 
@@ -94,10 +91,10 @@ class SimpleCircuitBreakerTest extends TestCase
      * @depends testCircuitBreakerIsInClosedStateAtStart
      * @depends testCircuitBreakerWillBeOpenedInCaseOfFailures
      * @depends testAfterTheThresholdTheCircuitBreakerMovesInHalfOpenState
+     * @dataProvider getCircuitBreakers
      */
-    public function testOnceInHalfOpenModeServiceIsFinallyReachable()
+    public function testOnceInHalfOpenModeServiceIsFinallyReachable($circuitBreaker)
     {
-        $circuitBreaker = $this->createCircuitBreaker();
         // CLOSED
         $circuitBreaker->call('https://httpbin.org/get/foo', $this->createFallbackResponse());
         // OPEN
@@ -112,7 +109,7 @@ class SimpleCircuitBreakerTest extends TestCase
                 $this->createFallbackResponse()
             )
         );
-        $this->assertSame(States::HALF_OPEN_STATE, $circuitBreaker->getState());
+        $this->assertSame('HALF_OPENED', $circuitBreaker->getState());
         $this->assertTrue($circuitBreaker->isHalfOpened());
 
         sleep(2);
@@ -125,20 +122,55 @@ class SimpleCircuitBreakerTest extends TestCase
             )
         );
 
-        $this->assertSame(States::CLOSED_STATE, $circuitBreaker->getState());
+        $this->assertSame('CLOSED', $circuitBreaker->getState());
         $this->assertTrue($circuitBreaker->isClosed());
+    }
+
+    /**
+     * Return the list of supported circuit breakers
+     *
+     * @return array
+     */
+    public function getCircuitBreakers()
+    {
+        return [
+            'simple' => $this->createSimpleCircuitBreaker(),
+            'symfony' => $this->createSymfonyCircuitBreaker(),
+        ];
     }
 
     /**
      * @return SimpleCircuitBreaker the circuit breaker for testing purposes
      */
-    private function createCircuitBreaker()
+    private function createSimpleCircuitBreaker()
     {
         return new SimpleCircuitBreaker(
             new OpenPlace(0, 0, 1), // threshold 1s
             new HalfOpenPlace(0, 0.2, 0), // timeout 0.2s to test the service
             new ClosedPlace(2, 0.2, 0), // 2 failures allowed, 0.2s timeout
             $this->getTestClient()
+        );
+    }
+
+    /**
+     * @return SymfonyCircuitBreaker the circuit breaker for testing purposes
+     */
+    private function createSymfonyCircuitBreaker()
+    {
+        $system = new MainSystem(
+            new ClosedPlace(2, 0.2, 0),
+            new HalfOpenPlace(0, 0.2, 0),
+            new OpenPlace(0, 0, 1)
+        );
+
+        $symfonyCache = new SymfonyCache(new ArrayCache());
+        $eventDispatcherS = $this->createMock(EventDispatcher::class);
+
+        return new SymfonyCircuitBreaker(
+            $system,
+            $this->getTestClient(),
+            $symfonyCache,
+            $eventDispatcherS
         );
     }
 
@@ -150,24 +182,5 @@ class SimpleCircuitBreakerTest extends TestCase
         return function () {
             return '{}';
         };
-    }
-
-    /**
-     * Returns an instance of Client able to emulate
-     * available and not available services.
-     *
-     * @return GuzzleClient
-     */
-    private function getTestClient()
-    {
-        $mock = new MockHandler([
-            new RequestException('Service unavailable', new Request('GET', 'test')),
-            new RequestException('Service unavailable', new Request('GET', 'test')),
-            new Response(200, [], '{"hello": "world"}'),
-        ]);
-
-        $handler = HandlerStack::create($mock);
-
-        return new GuzzleClient(['handler' => $handler]);
     }
 }
